teleport.c is the astral space implementation, and it is not one of the core concepts, so move it out of the kernel/ directory
